FIX: Ensure all images are converted to jpg on upload despite filename

+++ app/models/task.rb
@@ -781,13 +781,11 @@ class Task < ActiveRecord::Base
       # compress image files
       image_files = Dir.entries(task_dir).select { |f| (f =~ /^\d{3}.(image)/) == 0 }
       image_files.each do |img|
-        if File.extname(img) == ".jpg"
-          raise 'Failed to compress an image. Ensure all images are valid.' unless FileHelper.compress_image("#{task_dir}#{img}")
-        else
-          dest_file = "#{task_dir}#{File.basename(img, ".*")}.jpg"
-          raise 'Failed to compress an image. Ensure all images are valid.' unless FileHelper.compress_image_to_dest("#{task_dir}#{img}", dest_file, true)
-          FileUtils.rm("#{task_dir}#{img}")
-        end
+        # Ensure all images in submissions are not jpg
+        dest_file = "#{task_dir}#{File.basename(img, ".*")}.jpg"
+        raise 'Failed to compress an image. Ensure all images are valid.' unless FileHelper.compress_image_to_dest("#{task_dir}#{img}", dest_file, true)
+        # Cleanup unless the output was the same as the input
+        FileUtils.rm("#{task_dir}#{img}") unless dest_file == "#{task_dir}#{img}"
       end
 
       # copy all files into zip
